refactor strips_learning.py (#707)
* work in progress, largely untested

* some fixes

* progress but still wip

* more fixes and tests

* crapf

* minor

* sigh

* updates

* feedback
