+++ src/approaches/__init__.py
@@ -1,4 +1,5 @@
 """Handle creation of approaches."""
+
 import importlib
 import pkgutil
 from typing import TYPE_CHECKING, List, Set
