+++ src/nsrt_learning/nsrt_learning_main.py
@@ -11,10 +11,6 @@ from predicators.src import utils
 from predicators.src.nsrt_learning.option_learning import create_option_learner
 from predicators.src.nsrt_learning.sampler_learning import learn_samplers
 from predicators.src.nsrt_learning.segmentation import segment_trajectory
-from predicators.src.nsrt_learning.side_predicate_learning import \
-    BackchainingSidePredicateLearner, \
-    PredictionErrorHillClimbingSidePredicateLearner, \
-    PreserveSkeletonsHillClimbingSidePredicateLearner, SidePredicateLearner
 from predicators.src.nsrt_learning.strips_learning import \
     learn_strips_operators
 from predicators.src.settings import CFG
@@ -41,18 +37,15 @@ def learn_nsrts_from_data(trajectories: List[LowLevelTrajectory],
     segmented_trajs = [
         segment_trajectory(traj) for traj in ground_atom_dataset
     ]
-
-    # Each entry in ground_atom_dataset is a tuple of (low-level trajectory,
-    # low-level ground atoms sequence). We remove the latter, because
-    # it's prone to causing bugs -- we should rarely care about the
-    # low-level ground atoms sequence after segmentation.
-    low_level_trajs = [ll_traj for ll_traj, _ in ground_atom_dataset]
+    # We delete ground_atom_dataset because it's prone to causing bugs --
+    # we should rarely care about the low-level ground atoms sequence after
+    # segmentation.
     del ground_atom_dataset
 
     # If performing goal-conditioned sampler learning, we need to attach the
     # goals to the segments.
     if CFG.sampler_learning_use_goals:
-        for segment_traj, ll_traj in zip(segmented_trajs, low_level_trajs):
+        for segment_traj, ll_traj in zip(segmented_trajs, trajectories):
             # If the trajectory is not a demonstration, it does not have a
             # known goal (e.g., it was generated by random replay data), so we
             # won't attach a goal to the segment.
@@ -61,53 +54,25 @@ def learn_nsrts_from_data(trajectories: List[LowLevelTrajectory],
                 for segment in segment_traj:
                     segment.set_goal(goal)
 
-    segments = [seg for segs in segmented_trajs for seg in segs]
-
-    # STEP 3: Cluster the data by effects, jointly producing one STRIPSOperator,
-    #         Datastore, and OptionSpec per cluster. These items are then
-    #         used to initialize PartialNSRTAndDatastore objects (PNADs).
-    #         Note: The OptionSpecs here are extracted directly from the data.
-    #         If we are doing option learning, then the data will not contain
-    #         options, and so the option_spec fields are just the specs of a
-    #         DummyOption. We need a default dummy because future steps require
-    #         the option_spec field to be populated, even if just with a dummy.
+    # STEP 3: Learn STRIPS operators from the data, and use them to produce
+    #         PartialNSRTAndDatastore (PNAD) objects. Each PNAD contains a
+    #         STRIPSOperator, Datastore, and OptionSpec. The samplers will be
+    #         filled in on a later step.
     pnads = learn_strips_operators(
-        segments,
-        verbose=(CFG.option_learner != "no_learning"
-                 or CFG.side_predicate_learner != "no_learning"))
+        trajectories,
+        train_tasks,
+        predicates,
+        segmented_trajs,
+        verify_harmlessness=True,
+        verbose=(CFG.option_learner != "no_learning"))
 
-    # STEP 4: Learn side predicates for the operators and update PNADs. These
-    #         are predicates whose truth value becomes unknown (for *any*
-    #         grounding not explicitly in effects) upon operator application.
-    if CFG.side_predicate_learner != "no_learning":
-        assert CFG.option_learner == "no_learning", \
-            "Can't learn options and side predicates together."
-        if CFG.side_predicate_learner == "prediction_error_hill_climbing":
-            side_pred_learner: SidePredicateLearner = \
-                PredictionErrorHillClimbingSidePredicateLearner(
-                    pnads, trajectories, train_tasks, predicates,
-                    segmented_trajs)
-        elif CFG.side_predicate_learner == "preserve_skeletons_hill_climbing":
-            side_pred_learner = \
-                PreserveSkeletonsHillClimbingSidePredicateLearner(
-                    pnads, trajectories, train_tasks, predicates,
-                    segmented_trajs)
-        elif CFG.side_predicate_learner == "backchaining":
-            side_pred_learner = BackchainingSidePredicateLearner(
-                pnads, trajectories, train_tasks, predicates, segmented_trajs)
-        else:
-            raise ValueError(
-                f"side_predicate_learner {CFG.side_predicate_learner} not " +
-                "implemented")
-        pnads = side_pred_learner.sideline()
-
-    # STEP 5: Learn options (option_learning.py) and update PNADs.
+    # STEP 4: Learn options (option_learning.py) and update PNADs.
     _learn_pnad_options(pnads, action_space)  # in-place update
 
-    # STEP 6: Learn samplers (sampler_learning.py) and update PNADs.
+    # STEP 5: Learn samplers (sampler_learning.py) and update PNADs.
     _learn_pnad_samplers(pnads, sampler_learner)  # in-place update
 
-    # STEP 7: Log and return the NSRTs.
+    # STEP 6: Make, log, and return the NSRTs.
     nsrts = [pnad.make_nsrt() for pnad in pnads]
     logging.info("\nLearned NSRTs:")
     for nsrt in nsrts:
