+++ src/predicate_search_score_functions.py
@@ -142,8 +142,12 @@ class _OperatorLearningBasedScoreFunction(_PredicateSearchScoreFunction):
         # low-level ground atoms sequence after segmentation.
         low_level_trajs = [ll_traj for ll_traj, _ in pruned_atom_data]
         del pruned_atom_data
-        segments = [seg for traj in segmented_trajs for seg in traj]
-        pnads = learn_strips_operators(segments, verbose=False)
+        pnads = learn_strips_operators(low_level_trajs,
+                                       self._train_tasks,
+                                       set(candidate_predicates
+                                           | self._initial_predicates),
+                                       segmented_trajs,
+                                       verbose=False)
         strips_ops = [pnad.op for pnad in pnads]
         option_specs = [pnad.option_spec for pnad in pnads]
         op_score = self.evaluate_with_operators(candidate_predicates,
